Migrate website to Svelte + bun, add vibrational-modes demo, L-BFGS, CSVR
Website
- Rewrite from React to Svelte 5 (runes) + Vite 6 + bun
- Single SimulationStore ($state class) in context; components just read/bind
- Typed RPC wrapper around the worker (no postMessage switch-cases in UI)
- Decompose 1400-line MolecularDynamics.tsx into small components:
  ModelLoader, StructureLoader, Viewer, ViewerControls, RunControls,
  MDParams, OptParams, Stats, EnergyPlot, VibrationsPanel, Segmented,
  XyzEditorModal
- Pure chem utilities in lib/chem/ (bonds, sdf, cell, supercell, xyz)
- NGL isolated in lib/ngl/viewer.ts (imperative wrapper, orthographic camera,
  wide clip planes so small molecules don't get sliced on zoom)
- Layout: 4:3 viewer as hero, narrow side panels flanking; single-card frame
  housing viewer + plot strip; responsive down to phone
- Drag-and-drop model loader with green drop-feedback
- XYZ editor modal (Cmd+Enter to apply)
- Bundled model fetched from HuggingFace at prebuild time (curl → public/),
  not regenerated locally; gitignored

MD physics
- CSVR thermostat (Bussi-Donadio-Parrinello 2007) replaces Berendsen
- Maxwell-Boltzmann init rescaled to exact target T (fixes ~60% variance on
  small systems that made the thermostat look broken)
- Full atomic mass table (rows 1-5 + heavies); warns on unknown Z instead of
  silently defaulting to carbon
- Thermostat off (NVE) / forces type (conservative|NC) toggles exposed
- Energy drift diagnostic reported per mdStep
- Defaults: NVE + conservative forces (honest physics out of the box)

Optimization
- L-BFGS optimizer for atom-only paths with max-step cap (0.2 Å),
  backtracking Armijo line search, scaled-identity initial Hessian, m=10
  history, safety fallback to steepest descent on non-descent directions
- FIRE retained for cell+atom periodic optimization
- Routing: periodic + optimizeCell → FIRE, else L-BFGS; user-selectable
- 'optimizerStarted' event surfaces the active algorithm in the UI

Vibrational modes demo
- Worker 'predictAt' handler: predict at arbitrary positions reusing species
  and cell without touching MD state
- Jacobi eigensolver for symmetric real matrices (lib/vib/jacobi.ts)
- Finite-difference Hessian, symmetrized, mass-weighted, diagonalized
- Translation/rotation projector (standard OCC/ORCA recipe): build TR basis
  in mass-weighted coords, Gram-Schmidt with linear-molecule drop, sandwich
  P D P before diagonalization
- Frequencies in cm^-1 with imaginary modes flagged (negative freq)
- Auto-optimize-first toggle (default on) so modes are computed at a minimum
- Mode list with dominant-atom hint, click-to-animate along eigenvector,
  amplitude + period sliders
- Show/hide imaginary modes toggle

Worker bindings (src/api/wasm/mlipcpp_wasm.cpp)
- Return forces / stress / positions as Float32Array (avoid widening loop
  through embind val)
- Release previous model before loading a new one so old WebGPU device
  resources don't alias the new Predictor's buffers

Scripts
- scripts/export_pytorch/export_pet_full.py: _unwrap_to_pet walks common
  wrapper attributes (LLPRUncertaintyModel etc.) and falls back to scanning
  nn.Module children until a module with .gnn_layers is found
- scripts/convert_models.py: add pet-mad-xs to the default set
- scripts/publish_ggufs.py: new — push converted GGUFs to a HuggingFace repo
  (auto-creates LICENSE + README with BSD-3-Clause attribution, creates repo
  if missing, incremental re-upload)
- gguf/LICENSE + gguf/README.md committed as templates

CI
- .github/workflows/website.yml: switch to bun install --frozen-lockfile +
  bun run build, add setup-bun, drop uv (not needed now that prebuild is
  a curl), add GGUF cache keyed on package.json

+++ src/api/wasm/mlipcpp_wasm.cpp
@@ -92,9 +92,9 @@ public:
     bool isPeriodic() const { return periodic_; }
 
     val getPositions() const {
-        val result = val::global("Float64Array").new_(positions_.size());
+        val result = val::global("Float32Array").new_(positions_.size());
         for (size_t i = 0; i < positions_.size(); ++i) {
-            result.set(i, static_cast<double>(positions_[i]));
+            result.set(i, positions_[i]);
         }
         return result;
     }
@@ -111,9 +111,9 @@ public:
         if (!periodic_ || cell_.empty()) {
             return val::null();
         }
-        val result = val::global("Float64Array").new_(9);
+        val result = val::global("Float32Array").new_(9);
         for (int i = 0; i < 9; ++i) {
-            result.set(i, static_cast<double>(cell_[i]));
+            result.set(i, cell_[i]);
         }
         return result;
     }
@@ -248,18 +248,17 @@ public:
         val output = val::object();
         output.set("energy", static_cast<double>(result.energy));
 
-        // Convert forces to Float64Array
-        val forces = val::global("Float64Array").new_(result.forces.size());
+        // Return forces as Float32Array (native precision — no double-widening copy)
+        val forces = val::global("Float32Array").new_(result.forces.size());
         for (size_t i = 0; i < result.forces.size(); ++i) {
-            forces.set(i, static_cast<double>(result.forces[i]));
+            forces.set(i, result.forces[i]);
         }
         output.set("forces", forces);
 
-        // Include stress if available
         if (result.has_stress()) {
-            val stress = val::global("Float64Array").new_(6);
+            val stress = val::global("Float32Array").new_(6);
             for (int i = 0; i < 6; ++i) {
-                stress.set(i, static_cast<double>(result.stress[i]));
+                stress.set(i, result.stress[i]);
             }
             output.set("stress", stress);
         }
